requirements.txt
asgiref==3.2.7
awsebcli==3.18.1
beautifulsoup4==4.9.1
botocore==1.15.49
cement==2.8.2
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.6
django-allauth==0.41.0
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==1.1.1
django-crispy-forms==1.9.1
django-debug-toolbar==2.2
django-heroku==0.3.1
django-tinymce==3.0.2
django-tinymce4-lite==1.8.0
docutils==0.15.2
future==0.16.0
gunicorn==20.0.4
idna==2.7
jmespath==0.10.0
jsmin==2.2.2
oauthlib==3.1.0
pathspec==0.5.9
Pillow==7.1.2
psycopg2==2.8.5
pypiwin32==223
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2020.1
pywin32==227
PyYAML==5.3.1
requests==2.20.1
requests-oauthlib==1.3.0
semantic-version==2.5.0
six==1.11.0
soupsieve==2.0.1
sqlparse==0.3.1
termcolor==1.1.0
urllib3==1.24.3
wcwidth==0.1.9
whitenoise==5.0.1
